requirements.txt
# PhotoSense-AI Dependencies
# Python 3.10+ required

# Core dependencies
pillow>=10.0.0
numpy>=1.24.0

# Face detection and recognition
mtcnn>=0.1.1
tensorflow>=2.13.0
keras-facenet>=0.3.0

# Clustering
scikit-learn>=1.3.0

# Database
# SQLite is included in Python standard library

# Optional: Vector search (for future use)
# faiss-cpu>=1.7.4

# Optional: UI
streamlit>=1.28.0

# CLI
click>=8.1.0
